apps/crawler/requirements.txt
apache-airflow==2.9.2
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
SQLAlchemy==2.0.34
PyMySQL==1.1.1
python-dotenv==1.0.1
urllib3==2.2.2
